requirements.txt
---
# Core ML 
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# baseball data 
pybaseball>=2.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data processing
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# MLOps 
mlflow>=2.7.0

# Deployment (commented out for now, will revisit)
# fastapi>=0.104.0
# uvicorn>=0.24.0
# streamlit>=1.28.0
---
